requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
faiss-cpu
sentence-transformers
groq
python-dotenv